+++ ErsatzTV.Application/Scheduling/Queries/GetAllBlocksForPlayout.cs
@@ -1,0 +1,3 @@
+namespace ErsatzTV.Application.Scheduling;
+
+public record GetAllBlocksForPlayout(int PlayoutId) : IRequest<List<BlockViewModel>>;