+++ CHANGELOG.md
@@ -9,12 +9,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   - Add template data (like `MediaItem_Title`) for other video files
   - Add `MediaItem_Path` for movies, episodes, music videos and other videos
   - Add `get_directory_name` and `get_filename_without_extension` functions for path processing
+- Add `Block Playout Troubleshooting` tool to help investigate block playout history
 
 ### Fixed
 - Fix NVIDIA startup errors on arm64
 
 ### Changed
 - Do not use graphics engine for single, permanent watermark
+- Rename `YAML Validation` tool to `Sequential Schedule Validation`
 
 ## [25.7.1] - 2025-10-09
 ### Added